Добавил общий smart view
